git-svn-id: http://miranda.googlecode.com/svn/branches/stable@12085 83df23c2-51fd-11de-bb24-2d5e9b4af41d

+++ miranda/plugins/clist_nicer/CLUIFrames/cluiframes.c
@@ -3420,7 +3420,7 @@ static INT_PTR AddIconToExtraImageList(WPARAM wParam, LPARAM lParam)
 	if (himlExtraImages == 0 || wParam == 0)
 		return -1;
 
-	return((int)ImageList_AddIcon(himlExtraImages, (HICON)wParam));
+	return ImageList_AddIcon(himlExtraImages, (HICON)wParam);
 }
 
 /*
