+++ miranda/plugins/clist_nicer/clistsettings.c
@@ -30,6 +30,6 @@ extern int g_shutDown;
 INT_PTR GetContactStatusMessage(WPARAM wParam, LPARAM lParam)
 {
 	if(!g_shutDown)
-		return((int)SendMessage(pcli->hwndContactTree, CLM_GETSTATUSMSG, wParam, lParam));
+		return SendMessage(pcli->hwndContactTree, CLM_GETSTATUSMSG, wParam, lParam);
 	return 0;
 }
