+++ miranda/protocols/Gadu-Gadu/dynstuff.c
@@ -243,7 +243,7 @@ int string_append_n(string_t s, const char *str, int count)
 	}
 
 	if (count == -1)
-		count = strlen(str);
+		count = (int)strlen(str);
 
 	string_realloc(s, s->len + count);
 
@@ -276,7 +276,7 @@ void string_insert_n(string_t s, int index, const char *str, int count)
 		return;
 
 	if (count == -1)
-		count = strlen(str);
+		count = (int)strlen(str);
 
 	if (index > s->len)
 		index = s->len;
@@ -311,8 +311,8 @@ string_t string_init(const char *value)
 		value = "";
 
 	tmp->str = _strdup(value);
-	tmp->len = strlen(value);
-	tmp->size = strlen(value) + 1;
+	tmp->len = (int)strlen(value);
+	tmp->size = (int)strlen(value) + 1;
 
 	return tmp;
 }
